Optimse test_upgrade_path.py
