+++ tests/common/platform/templates/expect_boot_messages
