+++ tests/common/platform/warmboot_sad_cases.py
